+++ packages/builder/builder-rspack-provider/tests/plugins/css.test.ts
@@ -110,6 +110,23 @@ describe('plugins/css', () => {
       '"localIdentName":"[hash:5]"',
     );
   });
+
+  it('should use custom cssModules rule when using output.cssModules config', async () => {
+    const builder = await createBuilder({
+      plugins: [builderPluginCss()],
+      builderConfig: {
+        output: {
+          cssModules: {
+            auto: resourcePath => resourcePath.includes('.module.'),
+          },
+        },
+      },
+    });
+    const {
+      origin: { bundlerConfigs },
+    } = await builder.inspectConfig();
+    expect(bundlerConfigs[0]).toMatchSnapshot();
+  });
 });
 
 describe('plugins/less', () => {