+++ packages/builder/builder-rspack-provider/tests/plugins/default.test.ts
@@ -112,7 +112,6 @@ describe('bundlerApi', () => {
     expect(bundlerConfigs[0]).toMatchInlineSnapshot(`
       {
         "devtool": "hidden-source-map",
-        "module": {},
         "target": "node",
       }
     `);