+++ packages/builder/builder-shared/tests/utils.test.ts
@@ -1,5 +1,10 @@
 import { describe, expect, it } from 'vitest';
-import { awaitableGetter, isCssModules } from '../src';
+import {
+  awaitableGetter,
+  isCssModules,
+  getCssModulesAutoRule,
+  isLooseCssModules,
+} from '../src';
 
 describe('awaitableGetter', () => {
   it('should work', async () => {
@@ -36,4 +41,49 @@ it('check isCssModules', () => {
       },
     }),
   ).toBeFalsy();
+
+  expect(
+    isCssModules('src/index.module.css', {
+      auto: /\.module\./i,
+    }),
+  ).toBeTruthy();
+
+  expect(
+    isCssModules('src/index.css', {
+      auto: /\.module\./i,
+    }),
+  ).toBeFalsy();
+});
+
+it('getCssModulesAutoRule', () => {
+  expect(getCssModulesAutoRule(undefined, true)).toEqual(isLooseCssModules);
+
+  expect(
+    getCssModulesAutoRule(
+      {
+        auto: false,
+      },
+      true,
+    ),
+  ).toBeFalsy();
+
+  expect(
+    getCssModulesAutoRule(
+      {
+        auto: true,
+      },
+      true,
+    ),
+  ).toBeTruthy();
+
+  const autoFn = () => true;
+
+  expect(
+    getCssModulesAutoRule(
+      {
+        auto: autoFn,
+      },
+      true,
+    ),
+  ).toEqual(autoFn);
 });