+++ packages/cli/doc-core/src/node/plugins/autoNavAndSidebar/type.ts
@@ -5,8 +5,9 @@ export type NavMeta = NavItem[];
 export type SideMetaItem =
   | string
   | {
-      type: 'file' | 'dir';
+      type: 'file' | 'dir' | 'custom-link';
       name: string;
+      link?: string;
       // Use the h1 title as the sidebar title by default
       label?: string;
       collapsible?: boolean;