+++ tests/e2e/builder/cases/css/css-modules/src/index.js
@@ -1,7 +1,4 @@
 import './a.css';
-import b from './b.module.scss';
-import c from './c.module.less';
+import './b.module.scss';
+import './c.module.less';
 import './d.global.less';
-
-console.log(b);
-console.log(c);