+++ packages/cli/babel-preset-app/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @modern-js/babel-preset-app
 
+## 2.22.1
+
+### Patch Changes
+
+- Updated dependencies [e2848a2]
+- Updated dependencies [d4045ed]
+  - @modern-js/utils@2.22.1
+  - @modern-js/babel-preset-base@2.22.1
+
 ## 2.22.0
 
 ### Patch Changes