+++ packages/document/doc-tools-doc/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @modern-js/doc-tools-doc
 
+## 2.22.1
+
 ## 2.22.0
 
 ### Patch Changes