+++ packages/document/module-doc/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @modern-js/module-tools-docs
 
+## 2.22.1
+
+### Patch Changes
+
+- Updated dependencies [30a8d8d]
+- Updated dependencies [f6b2b3a]
+- Updated dependencies [5a5f164]
+  - @modern-js/doc-tools@2.22.1
+  - @modern-js/doc-plugin-auto-sidebar@2.22.1
+
 ## 2.22.0
 
 ### Minor Changes