+++ packages/generator/generators/generator-generator/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @modern-js/generator-generator
 
+## 3.1.21
+
 ## 3.1.20
 
 ### Patch Changes