+++ packages/review/eslint-config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @modern-js/eslint-config
 
+## 2.22.1
+
+### Patch Changes
+
+- @modern-js-app/eslint-config@2.22.1
+
 ## 2.22.0
 
 ### Patch Changes