+++ packages/runtime/plugin-garfish/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @modern-js/plugin-garfish
 
+## 2.22.1
+
+### Patch Changes
+
+- Updated dependencies [e2848a2]
+- Updated dependencies [9b69987]
+- Updated dependencies [d4045ed]
+  - @modern-js/runtime@2.22.1
+  - @modern-js/utils@2.22.1
+
 ## 2.22.0
 
 ### Patch Changes