+++ packages/server/plugin-koa/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @modern-js/plugin-koa
 
+## 2.22.1
+
+### Patch Changes
+
+- Updated dependencies [25b490a]
+- Updated dependencies [e2848a2]
+- Updated dependencies [d4045ed]
+  - @modern-js/types@2.22.1
+  - @modern-js/utils@2.22.1
+  - @modern-js/bff-core@2.22.1
+  - @modern-js/bff-runtime@2.22.1
+
 ## 2.22.0
 
 ### Patch Changes