+++ scripts/vitest-config/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @scripts/vitest-config
 
+## 2.22.1
+
+### Patch Changes
+
+- Updated dependencies [e2848a2]
+- Updated dependencies [d4045ed]
+  - @modern-js/utils@2.22.1
+  - @modern-js/tsconfig@2.22.1
+
 ## 2.22.0
 
 ### Patch Changes